* Giving some cache variables more sensible names 
* Removing unnecessary elements from the cache. 

Benchmark results: 

Before: 
```julia
[ Info: Benchmark: cache creation
[ Info: low order
  1.137 μs (16 allocations: 37.38 KiB)
[ Info: high order (q = 13)
  2.786 μs (18 allocations: 39.14 KiB)


[ Info: Benchnmark: function exection no cache / cache
[ Info: order = 3
  294.004 μs (71 allocations: 204.11 KiB)
  289.479 μs (54 allocations: 157.19 KiB)
[ Info: order = 5
  209.476 μs (59 allocations: 156.36 KiB)
  204.948 μs (42 allocations: 108.95 KiB)
[ Info: order = 7
  176.756 μs (53 allocations: 137.59 KiB)
  172.087 μs (36 allocations: 89.73 KiB)
[ Info: order = 9
  144.638 μs (47 allocations: 119.14 KiB)
  140.253 μs (30 allocations: 70.69 KiB)
[ Info: order = 13
  109.732 μs (37 allocations: 97.20 KiB)
  105.795 μs (18 allocations: 51.45 KiB)
```

After: 

```julia
[ Info: Benchmark: cache creation
[ Info: low order
  708.119 ns (10 allocations: 20.73 KiB)
[ Info: high order (q = 13)
  2.544 μs (16 allocations: 32.64 KiB)


[ Info: Benchnmark: function exection no cache / cache
[ Info: order = 3
  292.531 μs (65 allocations: 187.47 KiB)
  290.586 μs (54 allocations: 157.19 KiB)
[ Info: order = 5
  208.147 μs (53 allocations: 139.58 KiB)
  205.640 μs (42 allocations: 108.95 KiB)
[ Info: order = 7
  175.723 μs (47 allocations: 120.62 KiB)
  173.910 μs (36 allocations: 89.73 KiB)
[ Info: order = 9
  143.843 μs (41 allocations: 102.02 KiB)
  141.156 μs (30 allocations: 70.69 KiB)
[ Info: order = 13
  109.445 μs (35 allocations: 90.70 KiB)
  105.679 μs (18 allocations: 51.45 KiB)
```

Benchmark script:

```julia
using FiniteHorizonGramians, BenchmarkTools

λ = 1.0
n = 20
t = 1.0
A = λ * (I - 2 * tril(ones(n, n)))
B = sqrt(2λ) * ones(n, 1)
eA = similar(A)
U = similar(A)

qs = [3, 5, 7, 9, 13]

methods = [ExpAndGram{Float64,q}() for q in qs]
caches = [FiniteHorizonGramians.alloc_mem(A, B, method) for method in methods]

order(::ExpAndGram{T,Q}) where {T,Q} = Q

@info "Benchmark: cache creation"
@info "low order"
@btime FiniteHorizonGramians.alloc_mem($A, $B, ExpAndGram{Float64,3}())
@info "high order (q = 13)"
@btime FiniteHorizonGramians.alloc_mem($A, $B, ExpAndGram{Float64,13}())

println("\n")

@info "Benchnmark: function exection no cache / cache"
for (method, cache) in zip(methods, caches)
    @info "order = $(order(method))"
    @btime exp_and_gram_chol($A, $B, $t, $method)
    @btime exp_and_gram_chol!($eA, $U, $A, $B, $t, $method, $cache)
end

```